Add sub-headers to AWS and GCP sections
I've split each section into two components: the configuration required in the cloud provider; and the Terraform provider configuration. I've also reordered the AWS and GCP sections so that we do the cloud provider config first, and then Terraform. For AWS it doesn't matter, but for GCP and Azure we need inputs from the first step in order to configure the TF provider.