Ignore self-hosted doc links
The problem is that the self-hosted docs don't actually exist yet. Also, because of the way that Mike works, the branch containing the self-hosted releases will contain the built HTML rather than the markdown files. Since the number of links from the SaaS docs to self-hosted is likely to be limited, I think it's safe to ignore them for now.

In addition, I've updated the link pointing at the self-hosted section. I don't think we'll need the index.html part because of the way the CF distribution Lambda alters the URLs for us.