Add self-hosted release workflow
This new workflow allows us to tag the repo with a tag in the format `self-hosted-vx.y.z` to automatically publish a new version of the self-hosted documentation. This will run Mike, push the changes to the `self-hosted-releases` branch, and then trigger deployments.

For now I've left the production deploy commented out because I don't want to accidentally deploy to prod until we're happy.

I've also added concurrency groups to both the self-hosted release workflow and the main deployment workflow to avoid a situation where both attempt to deploy changes at the same time.