Build separate preprod/prod copies of self-hosted
The current version of the self-hosted docs build wasn't specifying the `DOC_ENV` flag. This meant that we were using the development configuration for things like analytics.

To solve what I've done is created two Mike builds - one for preprod and another for prod. They publish to two separate branches: `self-hosted-releases-preprod` and `self-hosted-releases-prod`. I haven't updated the main workflow yet to use those branches, and I've also disabled the deploy on the self-hosted-release workflow because we need the branches to be created before we can use them. I'll update the workflows in a separate PR.