Remove references to SaaS docs from Self-Hosted
I've adjusted some of the pages so that Self-Hosted pages aren't linking to sections that don't exist in the Self-Hosted docs. For now I've taken a pretty heavy handed approach of completely hiding the Azure and GCP examples for authentication with the Kubernetes integration, for example. I wanted to avoid including incorrect information rather than trying to adjust the examples to work without our managed integrations, and potentially misleading people.