Add info on providing admin consent for MS login
Users can be unable to sign up for a Spacelift account using their Azure AD account if their organization does not allow it. I've added some info on how to work around this.

At the moment the admin consent process results in an error message being displayed to users, but it actually works. I've included this in the docs to avoid confusion, but I'll raise a development ticket to check whether we can improve this.