Document change in precedence rules between config versions
We've added a v2 version of the config.yml file that brings the stack runtime config in line with the runtime config used for module tests cases, and grants stack defaults higher precedence than the current stack settings. This makes it possible to change settings across all stacks in one place, but is opt-in (via the version) for backwards compatibility.