Clarify that OIDC Federation requires access to Spacelift server instance
I've moved the explanation about needing the Spacelift server instance to be accessible by the service provider you are trying to connect to into its own section instead of a note at the top of the page. I've also detailed the URLs used, along with explaining that some providers may support manually uploading the keys.